requirements.txt
pandas
Flask
requests
beautifulsoup4
kafka-python
kafka-python-ng

